+++ embassy-mcxa/src/dma.rs
@@ -50,6 +50,8 @@
 //! transfer.await;
 //! ```
 
+#![allow(dead_code)]
+
 use core::future::Future;
 use core::marker::PhantomData;
 use core::pin::Pin;
@@ -60,7 +62,8 @@ use core::task::{Context, Poll};
 use embassy_hal_internal::{Peri, PeripheralType};
 use embassy_sync::waitqueue::AtomicWaker;
 
-use crate::clocks::Gate;
+use crate::clocks::enable_and_reset;
+use crate::clocks::periph_helpers::NoConfig;
 use crate::dma::sealed::SealedChannel;
 use crate::pac::dma::vals::Halt;
 use crate::pac::edma_0_tcd::regs::{TcdAttr, TcdBiterElinkno, TcdCiterElinkno, TcdCsr};
@@ -77,20 +80,17 @@ use crate::peripherals::DMA0;
 /// The function enables the DMA0 clock, releases reset, and configures the controller
 /// for normal operation with round-robin arbitration.
 pub(crate) fn init() {
-    unsafe {
-        // Enable DMA0 clock and release reset
-        DMA0::enable_clock();
-        DMA0::release_reset();
+    // Enable DMA0 clock and release reset
+    let _ = unsafe { enable_and_reset::<DMA0>(&NoConfig) };
 
-        // Configure DMA controller
-        pac::DMA0.mp_csr().modify(|w| {
-            w.set_edbg(true);
-            w.set_erca(true);
-            w.set_halt(Halt::NORMAL_OPERATION);
-            w.set_gclc(true);
-            w.set_gmrc(true);
-        });
-    }
+    // Configure DMA controller
+    pac::DMA0.mp_csr().modify(|w| {
+        w.set_edbg(true);
+        w.set_erca(true);
+        w.set_halt(Halt::NORMAL_OPERATION);
+        w.set_gclc(true);
+        w.set_gmrc(true);
+    });
 }
 
 /// DMA transfer priority.
@@ -243,6 +243,7 @@ pub const DMA_MAX_TRANSFER_SIZE: usize = 0x7FFF;
 #[derive(Clone, Copy, Debug)]
 #[repr(u8)]
 #[allow(dead_code)]
+#[cfg(feature = "mcxa2xx")]
 pub(crate) enum DmaRequest {
     WUU0WakeUpEvent = 1,
     CAN0 = 2,
@@ -295,11 +296,11 @@ pub(crate) enum DmaRequest {
     CMP1 = 54,
     CMP2 = 55,
     DAC0FifoRequest = 56,
-    GPIO0Pin = 60,
-    GPIO1Pin = 61,
-    GPIO2Pin = 62,
-    GPIO3Pin = 63,
-    GPIO4Pin = 64,
+    GPIO0PinEvent0 = 60,
+    GPIO1PinEvent0 = 61,
+    GPIO2PinEvent0 = 62,
+    GPIO3PinEvent0 = 63,
+    GPIO4PinEvent0 = 64,
     QDC0 = 65,
     QDC1 = 66,
     FlexIO0SR0 = 71,
@@ -322,6 +323,100 @@ pub(crate) enum DmaRequest {
     SGI0ReqOdat = 120,
     ADC2FifoRequest = 123,
     ADC3FifoRequest = 124,
+}
+
+/// DMA request sources
+///
+/// (from MCXA577 reference manual PDF attachment "DMA_Configuration.xml")
+#[derive(Clone, Copy, Debug)]
+#[repr(u8)]
+#[allow(dead_code)]
+#[cfg(feature = "mcxa5xx")]
+pub(crate) enum DmaRequest {
+    WUU0WakeUpEvent = 1,
+    CAN0 = 2,
+    LPI2C2Rx = 3,
+    LPI2C2Tx = 4,
+    LPI2C3Rx = 5,
+    LPI2C3Tx = 6,
+    I3C0Rx = 7,
+    I3C0Tx = 8,
+    I3C1Rx = 9,
+    I3C1Tx = 10,
+    LPI2C0Rx = 11,
+    LPI2C0Tx = 12,
+    LPI2C1Rx = 13,
+    LPI2C1Tx = 14,
+    LPSPI0Rx = 15,
+    LPSPI0Tx = 16,
+    LPSPI1Rx = 17,
+    LPSPI1Tx = 18,
+    LPSPI2Rx = 19,
+    LPSPI2Tx = 20,
+    LPUART0Rx = 21,
+    LPUART0Tx = 22,
+    LPUART1Rx = 23,
+    LPUART1Tx = 24,
+    LPUART2Rx = 25,
+    LPUART2Tx = 26,
+    LPUART3Rx = 27,
+    LPUART3Tx = 28,
+    LPUART4Rx = 29,
+    LPUART4Tx = 30,
+    Ctimer0M0 = 31,
+    Ctimer0M1 = 32,
+    Ctimer1M0 = 33,
+    Ctimer1M1 = 34,
+    Ctimer2M0 = 35,
+    Ctimer2M1 = 36,
+    Ctimer3M0 = 37,
+    Ctimer3M1 = 38,
+    Ctimer4M0 = 39,
+    Ctimer4M1 = 40,
+    LPTMR0CounterMatchEvent = 49,
+    ADC0FifoRequest = 51,
+    ADC1FifoRequest = 52,
+    CMP0 = 53,
+    DAC0FifoRequest = 56,
+    DAC1FifoRequest = 57,
+    GPIO5PinEvent0 = 59,
+    GPIO0PinEvent0 = 60,
+    GPIO1PinEvent0 = 61,
+    GPIO2PinEvent0 = 62,
+    GPIO3PinEvent0 = 63,
+    GPIO4PinEvent0 = 64,
+    TsiEndOfScan = 69,
+    TsiOutOfRange = 70,
+    FlexIO0SR0 = 71,
+    FlexIO0SR1 = 72,
+    FlexIO0SR2 = 73,
+    FlexIO0SR3 = 74,
+    CAN1 = 87,
+    EspiCh0 = 92,
+    EspiCh1 = 93,
+    LPI2C4Rx = 94,
+    LPI2C4Tx = 95,
+    LPSPI3Rx = 96,
+    LPSPI3Tx = 97,
+    LPSPI4Rx = 98,
+    LPSPI4Tx = 99,
+    LPUART5Rx = 100,
+    LPUART5Tx = 101,
+    I3C2Rx = 106,
+    I3C2Tx = 107,
+    I3C3Rx = 108,
+    I3C3Tx = 109,
+    FlexSPI0Rx = 110,
+    FlexSPI0Tx = 111,
+    ITRCTmprOut0 = 117,
+    SGI0ReqIdat = 119,
+    SGI0ReqOdat = 120,
+    Gpio0PinEvent1 = 132,
+    Gpio1PinEvent1 = 133,
+    Gpio2PinEvent1 = 134,
+    Gpio3PinEvent1 = 135,
+    Gpio4PinEvent1 = 136,
+    Gpio5PinEvent1 = 137,
 }
 
 impl DmaRequest {