+++ embassy-mcxa/src/gpio.rs
@@ -32,20 +32,14 @@ impl Iterator for BitIter {
     }
 }
 
+#[cfg(feature = "mcxa2xx")]
 const PORT_COUNT: usize = 5;
+#[cfg(feature = "mcxa5xx")]
+const PORT_COUNT: usize = 6;
 
-static PORT_WAIT_MAPS: [WaitMap<usize, ()>; PORT_COUNT] = [
-    WaitMap::new(),
-    WaitMap::new(),
-    WaitMap::new(),
-    WaitMap::new(),
-    WaitMap::new(),
-];
+static PORT_WAIT_MAPS: [WaitMap<usize, ()>; PORT_COUNT] = [const { WaitMap::new() }; PORT_COUNT];
 
 fn irq_handler(port_index: usize, gpio: crate::pac::gpio::Gpio, perf_wake: fn()) {
-    #[cfg(feature = "mcxa2xx")]
-    let isfr = gpio.isfr0();
-    #[cfg(feature = "mcxa5xx")]
     let isfr = gpio.isfr(0);
 
     for pin in BitIter(isfr.read().0) {
@@ -91,6 +85,13 @@ fn GPIO4() {
     irq_handler(4, crate::pac::GPIO4, crate::perf_counters::incr_interrupt_gpio4_wake);
 }
 
+#[cfg(feature = "mcxa5xx")]
+#[interrupt]
+fn GPIO5() {
+    crate::perf_counters::incr_interrupt_gpio5();
+    irq_handler(5, crate::pac::GPIO5, crate::perf_counters::incr_interrupt_gpio5_wake);
+}
+
 pub(crate) unsafe fn interrupt_init() {
     unsafe {
         use embassy_hal_internal::interrupt::InterruptExt;
@@ -100,6 +101,8 @@ pub(crate) unsafe fn interrupt_init() {
         crate::pac::interrupt::GPIO2.enable();
         crate::pac::interrupt::GPIO3.enable();
         crate::pac::interrupt::GPIO4.enable();
+        #[cfg(feature = "mcxa5xx")]
+        crate::pac::interrupt::GPIO5.enable();
 
         cortex_m::interrupt::enable();
     }
@@ -593,10 +596,6 @@ impl<'d> Flex<'d> {
         // Now that our waker is in the map, we can enable the appropriate interrupt
         //
         // Clear any existing pending interrupt on this pin
-        // TODO: Fix PAC naming
-        #[cfg(feature = "mcxa2xx")]
-        self.pin.gpio().isfr0().write(|w| w.0 = 1 << self.pin.pin());
-        #[cfg(feature = "mcxa5xx")]
         self.pin.gpio().isfr(0).write(|w| w.0 = 1 << self.pin.pin());
         self.pin
             .gpio()