+++ embassy-mcxa/src/perf_counters.rs
@@ -158,6 +158,8 @@ define_counters!(
     interrupt_gpio3_wake,
     interrupt_gpio4,
     interrupt_gpio4_wake,
+    interrupt_gpio5,
+    interrupt_gpio5_wake,
     interrupt_i2c0,
     interrupt_i2c0_wake,
     interrupt_i2c1,